Move request logics to the RequestHandler module
